refactor: update entry point and restructure main function in __main__.py
