+++ python/samples/gitty/src/gitty/__init__.py
@@ -1,0 +1,3 @@
+from .__main__ import main
+
+__all__ = ["main"]
